Cleaned up the code. Rearranged files
