+++ src/modules/localization/interface/dev_tools/obj_pose-laser-radar-synthetic-input.txt
